Changed constant data types to match their usage.
Fixes compiler warnings about type conversion data loss and comparing signed/unsigned.

+++ source/src/enet_encap/encap.h
@@ -22,7 +22,7 @@
 #define ENCAPSULATION_HEADER_LENGTH     24
 
 /** @brief Ethernet/IP standard port */
-static const int kOpenerEthernetPort = 0xAF12;
+static const unsigned short kOpenerEthernetPort = 0xAF12;
 
 /** @brief definition of status codes in encapsulation protocol
  * All other codes are either legacy codes, or reserved for future use
